requirements.txt
numpy
matplotlib
Flask
PyMySQL
python-telegram-bot
requests
SQLAlchemy>=1.2.7
pandas
pika
redis
pycparser
ratelimitingfilter
